@@ -10,18 +10,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^17.3.9",
-    "@angular/animations": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.4",
+    "@angular/animations": "^18.2.4",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
+    "@angular/common": "^18.2.4",
+    "@angular/compiler": "^18.2.4",
+    "@angular/core": "^18.2.4",
     "@angular/flex-layout": "^15.0.0-beta.42",
-    "@angular/forms": "^17.3.12",
+    "@angular/forms": "^18.2.4",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular/platform-browser": "^18.2.4",
+    "@angular/platform-browser-dynamic": "^18.2.4",
+    "@angular/router": "^18.2.4",
     "@gandlaf21/bolt11-decode": "^3.0.6",
     "@getalby/sdk": "^2.7.0",
     "@material/chips": "^15.0.0-canary.684e33d25.0",
@@ -45,8 +45,8 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular/cli": "~17.3.9",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/cli": "~18.2.4",
+    "@angular/compiler-cli": "^18.2.4",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
