@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HeaderComponent } from './components/header/header.component';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {APP_BASE_HREF} from '@angular/common';
 import { QRCodeModule } from 'angularx-qrcode';
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
@@ -159,96 +159,88 @@ const dbConfig: DBConfig  = {
     migrationFactory
 };
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    SidebarComponent,
-    FeedComponent,
-    PostComponent,
-    SettingsComponent,
-    LoginComponent,
-    RelayComponent,
-    CreateEventComponent,
-    UsersComponent,
-    MessengerComponent,
-    UserComponent,
-    UserDetailComponent,
-    Kind1Component,
-    ProfileEditComponent,
-    UsernamePipe,
-    SearchComponent,
-    NpubPipe,
-    HashtagPipe,
-    SafePipe,
-    TruncatePipe,
-    FollowingComponent,
-    PostDetailComponent,
-    NeventPipe,
-    EllipsisPipe,
-    ContactListComponent,
-    HumantimePipe,
-    FollowComponent,
-    LoadingComponent,
-    HashtagFeedComponent,
-    ZapComponent,
-    ImageDialogComponent,
-    NotificationsComponent,
-    NotificationComponent,
-    ListedUserComponent,
-    TrendingComponent,
-    ZapFeedComponent,
-    HomeFeedComponent,
-    UserBottomSheetComponent,
-    WalletComponent,
-    KeyboardPipe,
-    PaymentRequestComponent,
-    SendPaymentComponent,
-    PostQuickComponent,
-    HashtagQuickComponent,
-    MessagesListComponent,
-    PayInvoiceComponent,
-  ],
-  //entryComponents: [ImageDialogComponent, CreateEventComponent, UserBottomSheetComponent],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    HttpClientModule,
-    HttpCacheInterceptorModule.forRoot(),
-    MatIconModule,
-    MatCheckboxModule,
-    MatSliderModule,
-    MatToolbarModule,
-    MatMenuModule,
-    MatCardModule,
-    MatInputModule,
-    MatProgressSpinnerModule,
-    MatDialogModule,
-    MatChipsModule,
-    MatButtonModule,
-    MatBadgeModule,
-    MatDividerModule,
-    MatListModule,
-    MatExpansionModule,
-    MatTabsModule,
-    MatSelectModule,
-    MatIconModule,
-    MatProgressBarModule,
-    MatGridListModule,
-    FlexLayoutModule,
-    MatBottomSheetModule,
-    MatSnackBarModule,
-    AppRoutingModule,
-    QRCodeModule,
-    InfiniteScrollModule,
-    NgxIndexedDBModule.forRoot(dbConfig),
-    ReactiveFormsModule,
-    ZXingScannerModule,
-  ],
-  providers: [{provide: APP_BASE_HREF, useValue: '/'}],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        SidebarComponent,
+        FeedComponent,
+        PostComponent,
+        SettingsComponent,
+        LoginComponent,
+        RelayComponent,
+        CreateEventComponent,
+        UsersComponent,
+        MessengerComponent,
+        UserComponent,
+        UserDetailComponent,
+        Kind1Component,
+        ProfileEditComponent,
+        UsernamePipe,
+        SearchComponent,
+        NpubPipe,
+        HashtagPipe,
+        SafePipe,
+        TruncatePipe,
+        FollowingComponent,
+        PostDetailComponent,
+        NeventPipe,
+        EllipsisPipe,
+        ContactListComponent,
+        HumantimePipe,
+        FollowComponent,
+        LoadingComponent,
+        HashtagFeedComponent,
+        ZapComponent,
+        ImageDialogComponent,
+        NotificationsComponent,
+        NotificationComponent,
+        ListedUserComponent,
+        TrendingComponent,
+        ZapFeedComponent,
+        HomeFeedComponent,
+        UserBottomSheetComponent,
+        WalletComponent,
+        KeyboardPipe,
+        PaymentRequestComponent,
+        SendPaymentComponent,
+        PostQuickComponent,
+        HashtagQuickComponent,
+        MessagesListComponent,
+        PayInvoiceComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        HttpCacheInterceptorModule.forRoot(),
+        MatIconModule,
+        MatCheckboxModule,
+        MatSliderModule,
+        MatToolbarModule,
+        MatMenuModule,
+        MatCardModule,
+        MatInputModule,
+        MatProgressSpinnerModule,
+        MatDialogModule,
+        MatChipsModule,
+        MatButtonModule,
+        MatBadgeModule,
+        MatDividerModule,
+        MatListModule,
+        MatExpansionModule,
+        MatTabsModule,
+        MatSelectModule,
+        MatIconModule,
+        MatProgressBarModule,
+        MatGridListModule,
+        FlexLayoutModule,
+        MatBottomSheetModule,
+        MatSnackBarModule,
+        AppRoutingModule,
+        QRCodeModule,
+        InfiniteScrollModule,
+        NgxIndexedDBModule.forRoot(dbConfig),
+        ReactiveFormsModule,
+        ZXingScannerModule], providers: [{ provide: APP_BASE_HREF, useValue: '/' }, provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
 
 
